Delete a test involving Alt(5) (#34)
In GAP master we now use a generic character table for alternating
groups. As a result the content is now ordered differently. Both
versions are correct. To let the tests pass with both old and new
GAP, we just delete it -- there are other tests for this feature
right afterwards, and at least looking at them there seems to be
nothing unique about the test involving the alternating group. So
removing it seems reasonable.